+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## v2.7.0-alpha.3 (2020-08-25)
+
+#### :rocket: Enhancement
+
+- `linter`
+  - [#908](https://github.com/ampproject/amp-toolbox/pull/908) Linter rule for AMP Boilerplate removal ([@tharders](https://github.com/tharders))
+  - [#907](https://github.com/ampproject/amp-toolbox/pull/907) Linter rule for Google Font preconnect ([@tharders](https://github.com/tharders))
+
+#### Committers: 1
+
+- Thorsten Harders ([@tharders](https://github.com/tharders))
+
+## v2.7.0-alpha.2 (2020-08-21)
+
+#### :rocket: Enhancement
+
+- `linter`
+  - [#906](https://github.com/ampproject/amp-toolbox/pull/906) Add linter check for fast google fonts ([@tharders](https://github.com/tharders))
+  - [#904](https://github.com/ampproject/amp-toolbox/pull/904) Add linter check for preloading web fonts ([@tharders](https://github.com/tharders))
+  - [#902](https://github.com/ampproject/amp-toolbox/pull/902) Preload checks for blocking extensions ([@tharders](https://github.com/tharders))
+
+#### Committers: 2
+
+- Sebastian Benz ([@sebastianbenz](https://github.com/sebastianbenz))
+- Thorsten Harders ([@tharders](https://github.com/tharders))
+
 ## v2.7.0 (2020-08-19)
 
 #### :rocket: Enhancement